# src/frontend/openapi.json
{"openapi": "3.0.2", "info": {"title": "DanmakuIt Dashboard API", "description": "# \u7ba1\u7406\u9762\u677f\n\n\u201c\u5f39\u5e55\u4e00\u4e0b\u201d\u5e73\u53f0\u7684\u7ba1\u7406\u9762\u677fRESTful API.\n\n## Dependencies\n\n```shell\n$ pip install poetry\n$ poetry install # run with python 3.9\n```\n\n## \u767b\u5f55\n\n\u76ee\u524d\u5b9e\u73b0\u4e86\u901a\u8fc7\u5fae\u4fe1\u626b\u7801/GitHub/GitLab/TsinghuaGit\u8fdb\u884c\u7684\u793e\u4ea4\u7f51\u7edc\u767b\u5f55\uff1b\u5728\u626b\u7801\u540e\uff0c\u4f1a\u6839\u636e\u4ece\u5fae\u4fe1\u83b7\u53d6\u7684UnionID\u7b49\u4fe1\u606f\uff0c\n\u5efa\u7acb\u672c\u5730\u8d26\u6237\uff0c\u5e76\u4e14\u7b7e\u53d1\u76f8\u5e94\u7684 Json Web Token.\n\n## \u623f\u95f4\u7ba1\u7406\n\n\u76ee\u524d\u5b9e\u73b0\u4e86\u623f\u95f4\u7684\u589e\u52a0\u3001\u5217\u8868\u3001\u4fee\u6539\u3001\u5220\u9664\u7b49\u529f\u80fd\u3002\u548c\u5b9e\u65f6\u5f39\u5e55\u76f8\u5173\u7684\u529f\u80fd\u7531Apache Pulsar\u63d0\u4f9b\u3002\n\n## \u767b\u5f55\n\n\u767b\u9646\u540e Token \u5728\u524d\u7aef\u63a7\u5236\u53f0\u3002\n\n<!-- - <a target=\"_self\" href=\"/api/v1/user/social-login/github/login\">GitHub</a>\n- <a target=\"_self\" href=\"/api/v1/user/social-login/gitlab/login\">GitLab</a> -->\n<a target=\"_self\" href=\"/api/v1/user/social-login/gitlab3rd/login\">GitLabTsinghua Login</a>", "version": "0.1.0"}, "servers": [{"url": "https://danmakuit.panda2134.site/api/v1", "description": "Development server"}], "paths": {"/user/social-login/github/login": {"get": {"tags": ["user"], "summary": "Login Github", "operationId": "login_github_user_social_login_github_login_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/user/social-login/github/auth": {"get": {"tags": ["user"], "summary": "Auth Github", "operationId": "auth_github_user_social_login_github_auth_get", "responses": {"307": {"description": "redirect to frontend", "content": {"application/json": {"schema": {}}}}}}}, "/user/social-login/gitlab/login": {"get": {"tags": ["user"], "summary": "Login Gitlab", "operationId": "login_gitlab_user_social_login_gitlab_login_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/user/social-login/gitlab3rd/login": {"get": {"tags": ["user"], "summary": "Login Gitlab 3Rd Party", "operationId": "login_gitlab_3rd_party_user_social_login_gitlab3rd_login_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/user/social-login/gitlab/auth": {"get": {"tags": ["user"], "summary": "Auth Gitlab", "operationId": "auth_gitlab_user_social_login_gitlab_auth_get", "responses": {"307": {"description": "redirect to frontend", "content": {"application/json": {"schema": {}}}}}}}, "/user/social-login/gitlab3rd/auth": {"get": {"tags": ["user"], "summary": "Auth Gitlab 3Rd Party", "operationId": "auth_gitlab_3rd_party_user_social_login_gitlab3rd_auth_get", "responses": {"307": {"description": "redirect to frontend", "content": {"application/json": {"schema": {}}}}}}}, "/user/social-login/wechat": {"get": {"tags": ["user"], "summary": "Connect To Wechat", "operationId": "connect_to_wechat_user_social_login_wechat_get", "parameters": [{"required": true, "schema": {"title": "Code", "type": "string"}, "name": "code", "in": "query"}], "responses": {"307": {"description": "redirect to frontend", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/user/me": {"get": {"tags": ["user"], "summary": "User Information", "operationId": "user_information_user_me_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/User"}}}}}, "security": [{"HTTPBearer": []}]}}, "/room/": {"get": {"tags": ["room"], "summary": "List Room", "operationId": "list_room_room__get", "responses": {"200": {"description": "last 100 created rooms", "content": {"application/json": {"schema": {"title": "Response List Room Room  Get", "type": "array", "items": {"$ref": "#/components/schemas/Room"}}}}}}, "security": [{"HTTPBearer": []}]}, "post": {"tags": ["room"], "summary": "Create Room", "operationId": "create_room_room__post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/RoomNameModel"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Room"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"HTTPBearer": []}]}}, "/room/{room_id}": {"get": {"tags": ["room"], "summary": "Get Room", "operationId": "get_room_room__room_id__get", "parameters": [{"required": true, "schema": {"title": "Room Id", "type": "string"}, "name": "room_id", "in": "path"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Room"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"HTTPBearer": []}]}, "delete": {"tags": ["room"], "summary": "Delete Room", "operationId": "delete_room_room__room_id__delete", "parameters": [{"required": true, "schema": {"title": "Room Id", "type": "string"}, "name": "room_id", "in": "path"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RoomIdModel"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"HTTPBearer": []}]}, "patch": {"tags": ["room"], "summary": "Modify Room", "description": "uid, room_id and creation_time cannot be altered", "operationId": "modify_room_room__room_id__patch", "parameters": [{"required": true, "schema": {"title": "Room Id", "type": "string"}, "name": "room_id", "in": "path"}], "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/RoomUpdate"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Room"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"HTTPBearer": []}]}}, "/room/{room_id}/client-login": {"get": {"tags": ["room"], "summary": "Client Login Room", "description": "Set `room_passcode` in HTTP Bearer; `pulsar_jwt` is then used for pulsar connection", "operationId": "client_login_room_room__room_id__client_login_get", "parameters": [{"required": true, "schema": {"title": "Room Id", "type": "string"}, "name": "room_id", "in": "path"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Room"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"HTTPBearer": []}]}}, "/room/{room_id}/mpcode": {"get": {"tags": ["room"], "summary": "Get Room Mpcode", "description": "Get WeChat MiniProgram code for sending danmaku. Set `room_passcode` in HTTP Bearer;This is provided for clients so that they can fetch the QR code without JWT.", "operationId": "get_room_mpcode_room__room_id__mpcode_get", "parameters": [{"required": true, "schema": {"title": "Room Id", "type": "string"}, "name": "room_id", "in": "path"}], "responses": {"200": {"description": "respond with a JSON containing dataurl for MP Code.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RoomWeChatMpCodeResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"HTTPBearer": []}]}}, "/room/{room_id}/qrcode": {"get": {"tags": ["room"], "summary": "Get Room Qrcode", "description": "Set `room_passcode` in HTTP Bearer;This is provided for clients so that they can fetch the QR code without JWT.", "operationId": "get_room_qrcode_room__room_id__qrcode_get", "parameters": [{"required": true, "schema": {"title": "Room Id", "type": "string"}, "name": "room_id", "in": "path"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RoomQRCodeResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"HTTPBearer": []}]}}, "/room/{room_id}/fetch-subscribers": {"post": {"tags": ["room"], "summary": "Fetch Subscribers Of Room", "description": "Fetch the user information of all subscribers.Returns room_id in JSON when the fetch process starts.", "operationId": "fetch_subscribers_of_room_room__room_id__fetch_subscribers_post", "parameters": [{"required": true, "schema": {"title": "Room Id", "type": "string"}, "name": "room_id", "in": "path"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RoomIdModel"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"HTTPBearer": []}]}}, "/room/{room_id}/danmaku-admin": {"post": {"tags": ["room"], "summary": "Danmaku Admin Send", "description": "Send a danmaku message from admin. Sender in danmaku will always be overwritten to admin.", "operationId": "danmaku_admin_send_room__room_id__danmaku_admin_post", "parameters": [{"required": true, "schema": {"title": "Room Id", "type": "string"}, "name": "room_id", "in": "path"}], "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/DanmakuMessage"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DanmakuMessage"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"HTTPBearer": []}]}}, "/room/{room_id}/danmaku-update": {"post": {"tags": ["room"], "summary": "Danmaku Update", "description": "Update a danmaku message.", "operationId": "danmaku_update_room__room_id__danmaku_update_post", "parameters": [{"required": true, "schema": {"title": "Room Id", "type": "string"}, "name": "room_id", "in": "path"}], "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/DanmakuMessage"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DanmakuMessage"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"HTTPBearer": []}]}}, "/room/{room_id}/consumers": {"get": {"tags": ["room"], "summary": "Online Consumers", "description": "Get the online consumers of a room.", "operationId": "online_consumers_room__room_id__consumers_get", "parameters": [{"required": true, "schema": {"title": "Room Id", "type": "string"}, "name": "room_id", "in": "path"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"title": "Response Online Consumers Room  Room Id  Consumers Get", "type": "array", "items": {"$ref": "#/components/schemas/OnlineSubscription"}}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"HTTPBearer": []}]}}}, "components": {"schemas": {"ConsumerDetail": {"title": "ConsumerDetail", "type": "object", "properties": {"address": {"title": "Address", "type": "string"}, "consumerName": {"title": "Consumername", "type": "string"}, "availablePermits": {"title": "Availablepermits", "type": "integer"}, "blockedConsumerOnUnackedMsgs": {"title": "Blockedconsumeronunackedmsgs", "type": "boolean"}, "clientVersion": {"title": "Clientversion", "type": "string"}, "connectedSince": {"title": "Connectedsince", "type": "string"}, "msgRateOut": {"title": "Msgrateout", "type": "number"}, "msgRateRedeliver": {"title": "Msgrateredeliver", "type": "number"}, "msgThroughputOut": {"title": "Msgthroughputout", "type": "number"}, "unackedMessages": {"title": "Unackedmessages", "type": "integer"}}}, "DanmakuMessage": {"title": "DanmakuMessage", "required": ["color", "content", "id", "permission", "pos", "sender", "size"], "type": "object", "properties": {"color": {"title": "Color", "type": "string"}, "content": {"title": "Content", "type": "string"}, "id": {"title": "Id", "type": "string"}, "permission": {"title": "Permission", "type": "string"}, "pos": {"title": "Pos", "type": "string"}, "sender": {"title": "Sender", "type": "string"}, "size": {"title": "Size", "type": "string"}}}, "HTTPValidationError": {"title": "HTTPValidationError", "type": "object", "properties": {"detail": {"title": "Detail", "type": "array", "items": {"$ref": "#/components/schemas/ValidationError"}}}}, "OnlineSubscription": {"title": "OnlineSubscription", "required": ["subscription_name", "consumers"], "type": "object", "properties": {"subscription_name": {"title": "Subscription Name", "type": "string"}, "consumers": {"title": "Consumers", "type": "array", "items": {"$ref": "#/components/schemas/ConsumerDetail"}}}}, "Room": {"title": "Room", "required": ["name", "uid", "room_id", "room_passcode", "creation_time", "pulsar_jwt", "wechat_token"], "type": "object", "properties": {"name": {"title": "Name", "type": "string"}, "uid": {"title": "Uid", "type": "string"}, "danmaku_enabled": {"title": "Danmaku Enabled", "type": "boolean", "default": true}, "room_id": {"title": "Room Id", "type": "string"}, "room_passcode": {"title": "Room Passcode", "type": "string"}, "creation_time": {"title": "Creation Time", "type": "string", "format": "date-time"}, "remote_censor": {"title": "Remote Censor", "type": "boolean", "default": true}, "keyword_blacklist": {"title": "Keyword Blacklist", "type": "array", "items": {"type": "string"}, "default": []}, "pulsar_jwt": {"title": "Pulsar Jwt", "type": "string"}, "wechat_token": {"title": "Wechat Token", "type": "string"}, "wechat_encrypted": {"title": "Wechat Encrypted", "type": "boolean", "default": false}, "wechat_encryption_key": {"title": "Wechat Encryption Key", "type": "string", "default": ""}, "wechat_appid": {"title": "Wechat Appid", "type": "string", "default": ""}, "wechat_appsecret": {"title": "Wechat Appsecret", "type": "string", "default": ""}, "wechat_access_token": {"title": "Wechat Access Token", "type": "string", "default": ""}, "user_danmaku_colors": {"title": "User Danmaku Colors", "type": "array", "items": {"type": "string"}, "default": []}}}, "RoomIdModel": {"title": "RoomIdModel", "required": ["room_id"], "type": "object", "properties": {"room_id": {"title": "Room Id", "type": "string"}}}, "RoomNameModel": {"title": "RoomNameModel", "required": ["name"], "type": "object", "properties": {"name": {"title": "Name", "type": "string"}}}, "RoomQRCodeResponse": {"title": "RoomQRCodeResponse", "required": ["ticket", "expire_seconds", "url"], "type": "object", "properties": {"ticket": {"title": "Ticket", "type": "string"}, "expire_seconds": {"title": "Expire Seconds", "type": "integer"}, "url": {"title": "Url", "type": "string"}}}, "RoomUpdate": {"title": "RoomUpdate", "type": "object", "properties": {"name": {"title": "Name", "type": "string"}, "danmaku_enabled": {"title": "Danmaku Enabled", "type": "boolean", "default": true}, "remote_censor": {"title": "Remote Censor", "type": "boolean", "default": true}, "keyword_blacklist": {"title": "Keyword Blacklist", "type": "array", "items": {"type": "string"}, "default": []}, "wechat_token": {"title": "Wechat Token", "type": "string"}, "wechat_encrypted": {"title": "Wechat Encrypted", "type": "boolean", "default": false}, "wechat_encryption_key": {"title": "Wechat Encryption Key", "type": "string"}, "wechat_appid": {"title": "Wechat Appid", "type": "string"}, "wechat_appsecret": {"title": "Wechat Appsecret", "type": "string"}, "user_danmaku_colors": {"title": "User Danmaku Colors", "type": "array", "items": {"type": "string"}}}}, "RoomWeChatMpCodeResponse": {"title": "RoomWeChatMpCodeResponse", "required": ["image_dataurl"], "type": "object", "properties": {"image_dataurl": {"title": "Image Dataurl", "type": "string"}}}, "User": {"title": "User", "required": ["username", "uid"], "type": "object", "properties": {"username": {"title": "Username", "type": "string"}, "avatar": {"title": "Avatar", "type": "string"}, "uid": {"title": "Uid", "type": "string"}}}, "ValidationError": {"title": "ValidationError", "required": ["loc", "msg", "type"], "type": "object", "properties": {"loc": {"title": "Location", "type": "array", "items": {"type": "string"}}, "msg": {"title": "Message", "type": "string"}, "type": {"title": "Error Type", "type": "string"}}}}, "securitySchemes": {"HTTPBearer": {"type": "http", "scheme": "bearer"}}}}